Add files via upload
This project is a basic CRUD (Create, Read, Update, Delete) web application developed using JSP, Servlets, and MySQL. It demonstrates how to build a simple Java web application with proper MVC architecture, where JSP pages are used for the front-end, Servlets act as controllers, and JDBC is used to connect with the database. The application allows users to add new records, view all records in a table, update existing details, and delete records from the database. It is a beginner-friendly project to understand how JSP and Servlets work together for dynamic web development